+++ dddplus-test/src/test/java/io/github/dddplus/runtime/registry/mock/pattern/B2CPattern.java
@@ -5,8 +5,8 @@ import io.github.dddplus.ext.IIdentityResolver;
 import io.github.dddplus.model.IDomainModel;
 import io.github.dddplus.runtime.registry.mock.model.FooModel;
 
-@Pattern(code = BarPattern.CODE, name = "bar模式")
-public class BarPattern implements IIdentityResolver<IDomainModel> {
+@Pattern(code = B2CPattern.CODE, name = "B2C模式")
+public class B2CPattern implements IIdentityResolver<IDomainModel> {
     public static final String CODE = "bar";
 
     @Override