+++ RELEASE_NOTES.md
@@ -1,3 +1,17 @@
+## Release Notes - DDDplus - Version 1.1.0
+
+* FIXED
+   * 解决了`BaseDomainAbility`由于泛型机制不支持继承的问题
+   * [#7](https://github.com/dddplus/dddplus/issues/7) 解决`DomainArtifacts`暴露扩展点不全的问题：如果一个扩展点只被`Partner`实现，就没有暴露
+   * [#30](https://github.com/funkygao/cp-ddd-framework/issues/30) 解决了Plugin jar reload后，`DomainArtifacts`没有刷新的问题
+
+* Feature
+   * [#37](https://github.com/funkygao/cp-ddd-framework/issues/37) 除了`Partner/Pattern`的静态的扩展点路由机制外，提供了动态路由机制：`Policy`，供使用者扩展
+   * [#35](https://github.com/funkygao/cp-ddd-framework/issues/35) 提供绕过`BaseDomainAbility`而直接路由扩展点的机制，在业务属性不明显的场景下使用
+   * [#32](https://github.com/funkygao/cp-ddd-framework/issues/32) 框架提供默认的步骤编排能力，使用者不必从头编写`Ability`来编排步骤
+   * 增加了一个使用`DDDplus`来搭建`low-code`平台的例子工程：https://github.com/dddplus/easyapp
+      * 其中的Trigger机制依靠扩展点和Plugin的动态加载实现
+
 ## Release Notes - DDDplus - Version 1.0.3
 
 * FIXED